app/requirements.txt
python-dotenv
requests==2.32.4
urllib3==2.5.0
influxdb3-python>=0.14.0
prometheus_client>=0.19.0
certifi>=2025.8.3
charset-normalizer==3.4.2
idna==3.10
python-dateutil==2.9.0.post0
six==1.17.0
pydantic>=2.0.0
pydantic-settings>=2.0.3
PyYAML>=6.0
prometheus_client>=0.19.0
